refactor: align image viewer ui package
- keep new compose ui with existing compose code
